requirements.txt
Flask==2.3.3
gunicorn==21.2.0
tensorflow-cpu==2.10.0
Pillow==10.2.0
numpy==1.24.3

